Frontend not Linked but backend works
I have a mongodb cluster running online for ppl to access, it should be able to work without any more configurations as of rn. For this youll need to run npm install on both the frontend and the backend folder to be able to download all the dependencies. then u can npm start on the front and backend to get the website workin. im so bad at frontend that im having trouble geting it to communicate with the back, but the backend can send and recieve data from mongodb